refactor: kube utils
