--- pkg/client/pet.go
@@ -1,11 +1,0 @@
-package client
-
-import (
-	"fmt"
-
-	petname "github.com/dustinkirkland/golang-petname"
-)
-
-func Pet() string {
-	return fmt.Sprintf("chainsaw-%s", petname.Generate(2, "-"))
-}
