+++ pkg/utils/kube/unstructured.go
@@ -1,4 +1,4 @@
-package client
+package kube
 
 import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
