+++ pkg/utils/kube/unstructured_test.go
@@ -1,4 +1,4 @@
-package client
+package kube
 
 import (
 	"testing"
